requirements.txt
fastapi==0.109.2
pydantic==2.6.1
pydantic-settings==2.1.0
uvicorn==0.27.1
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7
email-validator==2.1.0.post1
celery==5.3.6
apscheduler==3.10.4
beautifulsoup4==4.12.3
selenium==4.17.2
streamlit==1.31.1
sendgrid==6.11.0
twilio==8.12.0
python-telegram-bot==20.7
python-dotenv==1.0.1
pandas==2.2.0
openpyxl==3.1.2
requests==2.31.0
plotly==5.18.0
numpy==1.26.4
python-dateutil==2.8.2
pytz==2024.1
jinja2==3.1.3
markupsafe==2.1.5
typing-extensions>=4.12.0,<5.0.0 